requirements.txt
requests
beautifulsoup4
discord-webhook
python-dotenv